🔧 chore(#33): Update CORS and FastAPI settings
